Move RPC stuff from NotebookUtils to RPCUtils
